lotto-saju-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
supabase==2.0.0
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
manse==0.1.0